+++ lib/backend/backend.dart
@@ -4,6 +4,7 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:trashtrace/data/models/dustbin.dart';
 import 'package:http/http.dart' as http;
+import 'package:trashtrace/data/models/recyclexmodel.dart';
 
 class ResponseType<T> {
   T result;
@@ -92,6 +93,30 @@ class TrashTraceBackend {
     );
   }
 
+  Future<ResponseType<int?>> getUserID({required String username}) async {
+    final root = await getServerLink();
+    final uri = Uri.parse("$root/user/get_id_by_username/$username");
+    final res = await http.get(uri);
+    if (res.statusCode == 200) {
+      final resbody = jsonDecode(res.body);
+      final id = int.parse(resbody['id'].toString());
+      return ResponseType(result: id, message: 'success');
+    }
+    return ResponseType(result: null, message: res.body);
+  }
+
+  Future<ResponseType<double?>> getUserPoints({required int userID}) async {
+    final root = await getServerLink();
+    final uri = Uri.parse("$root/user/get_user_points/$userID");
+    final res = await http.get(uri);
+    if (res.statusCode == 200) {
+      final resbody = jsonDecode(res.body);
+      final id = double.parse(resbody['points'].toString());
+      return ResponseType(result: id, message: 'success');
+    }
+    return ResponseType(result: null, message: res.body);
+  }
+
   Future<ResponseType<List<Dustbin>?>> getAllBins() async {
     final root = await getServerLink();
 
@@ -162,17 +187,18 @@ class TrashTraceBackend {
     return ResponseType(result: partners, message: 'success');
   }
 
-  Future<ResponseType<List<Map>?>> getAllMyJobs(int userId) async {
+  Future<ResponseType<List<ReCyclXJob>?>> getAllMyJobs(int userId) async {
     final root = await getServerLink();
 
-    List<Map> jobs = [];
+    List<ReCyclXJob> jobs = [];
     final uri = Uri.parse("$root/recyclx/myjobs/$userId");
     final res = await http.get(uri);
 
     if (res.statusCode == 200) {
       final resdata = jsonDecode(res.body);
       for (final x in resdata) {
-        jobs.add(x);
+        final rcxj = ReCyclXJob.fromMap(x);
+        jobs.add(rcxj);
       }
     } else {
       return ResponseType(
@@ -201,10 +227,12 @@ class TrashTraceBackend {
         {
           'name': name,
           'status': status,
-          'lat': loc.latitude,
-          'lng': loc.longitude,
-          'uid': userId,
-          'pid': partnerId
+          'destination': {
+            'lat': loc.latitude,
+            'lng': loc.longitude,
+          },
+          'user_id': userId,
+          'partner_id': partnerId
         },
       ),
     );
@@ -220,5 +248,29 @@ class TrashTraceBackend {
     }
     print('ERROR => ${res.body}');
     return ResponseType(result: null, message: res.body);
+  }
+
+  // =================== TrashTag ====================
+
+  Future<ResponseType<bool>> add2dustbin({
+    required int userId,
+    required String qrCodeValue,
+  }) async {
+    final root = await getServerLink();
+    final uri = Uri.parse("$root/trashtag/userscan");
+    final res = await http.post(
+      uri,
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: jsonEncode(
+        {'uid': userId, 'qrCodeValue': qrCodeValue},
+      ),
+    );
+    if (res.statusCode == 200) {
+      return ResponseType(result: true, message: res.body);
+    }
+    print('ERROR => ${res.body}');
+    return ResponseType(result: false, message: res.body);
   }
 }