+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 11.10.2023
+- Manas:
+    - TrashTag: Added Working Logic to Scan and Add Item to Dusbin & Show updated points!
+    - ReCyclX: Added the MyJobs Section & Job Booking Logic
+    - Backend: Added the Latest Routes
+
 # 22.08.2023
 - Manas:
     - Changed to Permalink Vercel Hosted Backend as fallback